requirements.txt
attrs==23.2.0
blis==0.2.4
catalogue==1.0.2
certifi==2024.2.2
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
cymem==2.0.8
docx2txt==0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.1.0/en_core_web_sm-2.1.0.tar.gz#sha256=4db16860a8cdef56d436038ace6abeb9181a5176bdc8c16c755a20dced51e5f1
et-xmlfile==1.1.0
idna==3.6
importlib-metadata==6.7.0
importlib-resources==5.12.0
joblib==1.3.2
jsonschema==3.0.1
murmurhash==1.0.10
nltk==3.8.1
numpy==1.21.6
openpyxl==3.1.2
pandas==1.3.5
pdfminer.six==20231228
pkgutil_resolve_name==1.3.10
plac==0.9.6
preshed==2.0.1
pycparser==2.21
pycryptodome==3.20.0
pyresparser==1.0.6
pyrsistent==0.19.3
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
sortedcontainers==2.4.0
spacy==2.1.4
srsly==1.0.7
thinc==7.0.8
tqdm==4.66.2
typing_extensions==4.7.1
urllib3==2.0.7
wasabi==0.10.1
zipp==3.15.0
